+++ components/urlRules/UrlRuleContentElement.php
@@ -7,7 +7,6 @@
  */
 namespace skeeks\cms\components\urlRules;
 use skeeks\cms\App;
-use skeeks\cms\filters\NormalizeDir;
 use skeeks\cms\models\CmsContentElement;
 use skeeks\cms\models\Tree;
 use \yii\base\InvalidConfigException;
@@ -138,8 +137,7 @@ class UrlRuleContentElement
      */
     protected function _normalizeDir($pathInfo)
     {
-        $filter             = new NormalizeDir();
-        $pathInfoNormal     = $filter->filter($pathInfo);
+        $pathInfoNormal     = $this->_filterNormalizeDir($pathInfo);
 
         if ((bool) \Yii::$app->seo->useLastDelimetrContentElements)
         {
@@ -148,6 +146,25 @@ class UrlRuleContentElement
         {
             return $pathInfoNormal;
         }
+    }
+
+    /**
+     * @param string $dir
+     * @return string
+     */
+    protected function _filterNormalizeDir($dir)
+    {
+        $result = [];
+
+        $data = explode(DIRECTORY_SEPARATOR, $dir);
+        foreach ($data as $value)
+        {
+            if ($value)
+            {
+                $result[] = $value;
+            }
+        }
+        return implode(DIRECTORY_SEPARATOR, $result);
     }
 
 }